move utilities path to outside of the e2e folder because it is not a test file
