+++ cypress/support/e2e.js
@@ -1,7 +1,7 @@
 /* global cy, Cypress */
 
 // ***********************************************************
-// This example support/index.js is processed and
+// This example support/e2e.js is processed and
 // loaded automatically before your test files.
 //
 // This is a great place to put global configuration and
